Chaged the decorator pattern to have 2 diferent usages, decrate and extend.
Decorate is aimed at injecting prpeties and methods into existin objects, which is in esence it's main usage.

Extend is used to inherit decorate, extend and all other properties and methods between prototypes.

Decorates are no longer instantiable on there own, eiter extend a new prototype function with the extend method or decorae an existing object with the decorate method